finished

+++ graph.js
@@ -10,6 +10,14 @@ Array.prototype.contains = function(name) {
     return false;
 }
 
+Object.arraySize = function(obj) {
+    var size = 0, key;
+    for (key in obj){
+        if(obj.hasOwnProperty(key)) size ++;
+    }
+    return size;
+}
+
 var Graph = function() {
     this.node_list = [];
 }
@@ -22,7 +30,7 @@ Graph.prototype.addEdge = function ( start, end, relationship = "") {
         //get start node
         var i = this.node_list.length;
         while (i--) {
-            if (this.node_list[i].name === start) {
+            if (this.node_list[i] && this.node_list[i].name === start) {
                 this.node_list[i].addEdge(end, relationship);
                 break;
             }
@@ -44,9 +52,9 @@ Graph.prototype.addEdge = function ( start, end, relationship = "") {
 
 Graph.prototype.addMultEdge = function( objArray ){ // obj = ['start', 'end', 'relationship']
 
-    for(var  i = 0 ; i < objArray; i ++){
-
-        objArray[i] = this.addEdge(objArray[i][0], objArray[i][1], objArray[i][2]);
+    for(var  i = 0 ; i < objArray.length; i ++){
+        if(objArray[i] && objArray[i][0] && objArray[i][1] && objArray[i][2])
+            this.addEdge(objArray[i][0], objArray[i][1], objArray[i][2]);
 
     }
 
@@ -61,9 +69,34 @@ Graph.prototype.deleteEdge = function( start, end, relationship = '' ){
         var i = this.node_list.length;
         while (i-- && i >= 0 ) {
 
-            if (this.node_list[i].name === start) {
+            if (this.node_list[i] && this.node_list[i].name === start) {
                 this.node_list[i].deleteEdge(end, relationship);
             }
+
+        }
+    }
+
+}
+
+Graph.prototype.deleteAllEdge = function( start, end, directional = true ){
+
+    var first = this.node_list.contains(start);
+    var second = this.node_list.contains(end);
+
+    if ( first && second ) {
+        var i = this.node_list.length;
+        while (i-- && i >= 0 ) {
+
+            if (this.node_list[i] && this.node_list[i].name === start) {
+                this.node_list[i].deleteAllEdge(end);
+            }
+
+            if (!directional) {
+                if (this.node_list[i] && this.node_list[i].name === end) {
+                    this.node_list[i].deleteAllEdge(start);
+                }
+            }
+
 
         }
     }
@@ -77,10 +110,13 @@ Graph.prototype.deleteNode = function( node ) {
     if ( first ) {
         var i = this.node_list.length;
         while ( i-- && i >= 0 ) {
-            if ( this.node_list[i].name === node ){
-                delete this.node_list[i];
-            } else {
-                this.node_list[i].deleteAllEdge( node );
+
+            if ( this.node_list[i] ) {
+                if ( this.node_list[i].name === node ){
+                    delete this.node_list[i];
+                } else {
+                    this.node_list[i].deleteAllEdge( node );
+                }
             }
 
         }
@@ -88,8 +124,32 @@ Graph.prototype.deleteNode = function( node ) {
 
 }
 
-Graph.prototype.clean = function( node ) {
-    //TODO
+Graph.prototype.clean = function() {
+    // delete node where no connections connect to it
+    for(var i = 0; i < this.node_list.length; i++){
+
+        if( this.node_list[i] && !Object.arraySize(this.node_list[i].edge_list) ){ //not a subject_node
+            var exists_relationship = false;
+            for(var j = 0; j < this.node_list.length; j ++){
+
+                if ( j != i && this.node_list[j] ) {
+                    for(var key in this.node_list[j].edge_list){
+                        if( key === this.node_list[i].name ) {// exists_relationship with this node
+                            exists_relationship = true;
+                            break;
+                        }
+                    }
+                }
+
+                if(exists_relationship) break; // found relationship
+
+            }
+            if(!exists_relationship){
+                console.log('delete here = ' + this.node_list[i].name);
+                delete this.node_list[i];
+            }
+        }
+    }
 }
 
 Graph.prototype.printNodes = function (){
